pyproject.toml
[tool.poetry]
name = "steel-imports"
version = "0.1.0"
description = ""
authors = ["Philipp Verpoort <philippv@pik-potsdam.de>"]

[tool.poetry.dependencies]
python = "^3.9"
plotly = "^5.4.0"
dash = "^2.0.0"
dash-bootstrap-components = "^1.0.1"
pandas = "^1.3.4"
openpyxl = "^3.0.9"
PyYAML = "^6.0"
kaleido = "0.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
